+++ preppal-fe/src/app.tsx
@@ -5,11 +5,11 @@ import Home from './pages/home';
 import LoginPage from './pages/login';
 import Signup from './pages/signup';
 import Profile from './pages/profile';
-import NewRecipe from './pages/new-recipe';
 import Collections from './pages/collections';
 import EditProfile from './pages/edit-profile';
 import Search from './pages/search';
 import Recipe from './pages/recipe';
+import CreateRecipe from './pages/create-recipe';
 
 function App() {
   return (
@@ -21,7 +21,7 @@ function App() {
           <Route path="signup" element={<Signup />} />
           <Route path="profile" element={<Profile />} />
           <Route path="edit-profile" element={<EditProfile />} />
-          <Route path="new-recipe" element={<NewRecipe />} />
+          <Route path="new-recipe" element={<CreateRecipe />} />
           <Route path="collections" element={<Collections />} />
           <Route path="search/:query" element={<Search />} />
           <Route path=":author/:titleUrl" element={<Recipe />} />
